+++ src/main/java/com/empatkepala/entity/request/AddFpkRequest.java
@@ -1,7 +1,18 @@
 package com.empatkepala.entity.request;
 
+import java.util.Date;
+
 /**
  * Created by ALz on 3/11/2017.
  */
 public class AddFpkRequest {
+    private int position;
+    private Date date;
+    private String reason;
+    private String fitnessWithMpp;
+    private String employeeStatus;
+    private String school;
+    private String workExperience;
+    private String skillKnowledge;
+    private boolean completeness;
 }
